% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.r
\name{plot_matrix}
\alias{plot_matrix}
\title{plot_matrix}
\usage{
plot_matrix(
  y,
  standardize.rows = TRUE,
  reorder.rows = TRUE,
  reorder.cols = TRUE,
  high.contrast = TRUE
)
}
\arguments{
\item{y}{data matrix to be evaluated.}

\item{standardize.rows}{logical argument (default = TRUE) indicating whether to standardize 
the rows of y to build the image.}

\item{reorder.rows}{logical argument (default = TRUE) indicating whether to reorder the rows 
of y to highlight a pattern.}

\item{reorder.cols}{logical argument (default = TRUE) indicating whether to reorder the columns 
of y to highlight a pattern.}

\item{high.contrast}{logical argument (default = TRUE) indicating whether to apply a transformation 
to increase contrast in the image of y.}
}
\description{
Function to build a heat map displaying the values of given data matrix. This graph is useful for a visual inspection of the 
spatial distribution of the observations within the target matrix.
}
\seealso{
\code{\link{slfm}}, \code{\link{process_matrix}}, \code{\link{slfm_list}}
}
