% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slfm.r
\name{print.slfm}
\alias{print.slfm}
\title{Print method for an slfm object}
\usage{
\method{print}{slfm}(x, ...)
}
\arguments{
\item{x}{object of the class 'slfm'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for an slfm object
}
