\name{process_matrix}
\alias{process_matrix}
\title{Pre-process data for SLFM}
\usage{
process_matrix(path, output_path, sample_size)
}
\arguments{
  \item{list}{list of matrices with the pre-processed data}
}
\description{
This function pre-process the data to be used for fitting a
sparse latent factor model.
}

