\docType{package}
\name{slfm-package}
\alias{slfm-package}
\title{slfm: the sparse latent factor model package for R.}
\description{
slfm is a set of tools to fit sparse latent factor models
to microarray data. This includes functions to:
}
\details{
* pre-process a set of matrices * fit models to a set of
matrices * detailed summary of model fit

Considerable effort has been put into making slfm fast and
memory efficient, so as slfm is an attractive alternative
to simpler methods in terms of execution time.
}

