% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.r
\name{plot_matrix}
\alias{plot_matrix}
\title{Image graph displaying a given data matrix.}
\usage{
plot_matrix(y, standardize.rows = TRUE, reorder.rows = TRUE,
  reorder.cols = TRUE, high.contrast = TRUE)
}
\arguments{
\item{y}{data matrix to be evaluated.}

\item{standardize.rows}{logical argument (default = TRUE) indicating whether to standardize 
the rows of y to build the image.}

\item{reorder.rows}{logical argument (default = TRUE) indicating whether to reorder the rows 
of y to highlight a pattern.}

\item{reorder.cols}{logical argument (default = TRUE) indicating whether to reorder the columns 
of y to highlight a pattern.}

\item{high.contrast}{logical argument (default = TRUE) indicating whether to apply a transformation 
to increase contrast in the image of y.}
}
\description{
This function builds an image graph displaying a heat map of a given data matrix.
}
