% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_matrix.r
\name{plot_matrix}
\alias{plot_matrix}
\title{Plot matrix}
\usage{
plot_matrix(y, standardize.rows = TRUE, reorder.rows = TRUE,
  reorder.cols = TRUE, high.contrast = TRUE)
}
\arguments{
\item{y}{matrix to be ploted}

\item{standardize.rows}{standardize matrix rows for plot}

\item{reorder.rows}{reorder matrix rows based on pattern}

\item{reorder.cols}{reorder matrix cols based on pattern}

\item{high.contrast}{apply transformation to matrix to increase contrast}
}
\description{
This function plots a data matrix
}

