% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_matrix.r
\name{process_matrix}
\alias{process_matrix}
\title{process_matrix}
\usage{
process_matrix(path, output_path, sample_size)
}
\arguments{
\item{path}{path to the directory containing the set of matrices to be pre-processed.}

\item{output_path}{path to the directory intended to acommodate the saved pre-processed matrices.}

\item{sample_size}{number of matrices to be used on the principal component analysis.}
}
\description{
Function to apply a procedure to pre-process the gene expression data saved in matrices. This pre-processing step 
is required to allow a fair analysis of the results obtained from the SLFM model.
}
\seealso{
\code{\link{slfm}}, \code{\link{slfm_list}}, \code{\link{plot_matrix}}
}
