% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_matrix.r
\name{process_matrix}
\alias{process_matrix}
\title{Procedure to pre-process the gene expression data to be fitted via SLFM.}
\usage{
process_matrix(path, output_path, sample_size)
}
\arguments{
\item{path}{path to the directory containing the set of matrices to be pre-processed.}

\item{output_path}{path to the directory intended to acommodate the saved pre-processed matrices.}

\item{sample_size}{number of matrices to be used on the principal component analysis.}
}
\description{
This function is intended to pre-process the data matrices to be fitted 
through the Sparse Latent Factor Model.
}
