\name{process_matrix}
\alias{process_matrix}
\title{Pre-process data for SLFM}
\usage{
process_matrix(path, output_path, sample_size)
}
\arguments{
  \item{path}{path containing the set of matrices to be
  processed}

  \item{output_path}{path to save the processed matrices}

  \item{sample_size}{number of matrices to be used on the
  principal component analysis}
}
\description{
This function pre-process the data to be used for fitting a
sparse latent factor model.
}

